Revert "Move tethering runtime out of app-fabric"
This reverts commit ed8a64908be1507618e6e4195e0849540e6abdb7.
