+++ cdap-app-fabric/src/test/java/io/cdap/cdap/internal/tethering/runtime/spi/provisioner/TetheringProvisionerTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package io.cdap.cdap.internal.tethering.runtime.spi.provisioner;
+
+import com.google.inject.Guice;
+import com.google.inject.Injector;
+import io.cdap.cdap.common.conf.CConfiguration;
+import io.cdap.cdap.internal.guice.AppFabricTestModule;
+import io.cdap.cdap.internal.provision.ProvisioningService;
+import io.cdap.cdap.messaging.MessagingService;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TetheringProvisionerTest {
+
+  @Test
+  public void testGuiceInjector() {
+    Injector injector = Guice.createInjector(new AppFabricTestModule(CConfiguration.create()));
+    Assert.assertNotNull(injector.getInstance(ProvisioningService.class));
+    Assert.assertNotNull(injector.getInstance(MessagingService.class));
+    Assert.assertNotNull(injector.getInstance(TetheringProvisioner.class));
+  }
+}
