--- cdap-runtime-ext-tethering/.gitignore
@@ -1,1 +1,0 @@
-data/
